\docType{data}
\name{debt}
\alias{debt}
\title{Data related to the Reinhart and Rogoff study of debt and GDP growth}
\format{data frame with 1171 rows and 4 variables}
\usage{
debt
}
\description{
The data is taken from Cosma Shalizi's UDA course:
\url{http://www.stat.cmu.edu/~cshalizi/uADA/13/hw/11/solutions-11.pdf}
}
\details{
\itemize{ \item Country. country name \item Year. year of
observation \item growth. real GDP growth rate \item ratio.
ratio of debt to GDP }
}
\references{
Carmen M. Reinhart and Kenneth S. Rogoff, "Growth in a Time
of Debt", American Economic Review 100 (2010): 573-578.
Thomas Herndon, Michael Ash and Robert Pollin, "Does High
Public Debt Consistently Stifle Economic Growth? A Critique
of Reinhart and Rogoff", University of
Massachusets-Amherst, Working Paper 332, 2013.
\url{http://www.peri.umass.edu/236/hash/31e2ff374b6377b2ddec04deaa6388b1/publication/566/}
}
\keyword{datasets}

