% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DropNA.psData}
\alias{DropNA.psData}
\title{Drop rows from a data frame with missing values in the OutCountryID variable.}
\source{
Largely based on \code{DropNA} from the \code{DataCombine} package.
}
\usage{
DropNA.psData(data, countryVar = "country", timeVar = NULL, OutCountryID)
}
\arguments{
\item{data}{a data frame object.}

\item{countryVar}{a character vector naming the variables you would like to
have only non-missing (NA) values.}

\item{timeVar}{variable indicating the time.}

\item{OutCountryID}{the ID the output country identifier.}
}
\description{
Drop rows from a data frame with missing values in the OutCountryID variable.
}
\keyword{internals}

