\name{try_date}
\alias{try_date}
\title{Autodetect a date format}
\usage{
try_date(x, date = "\%Y")
}
\value{
the fraction of perfect matches
}
\description{
Looks for perfect matches between the values of a variable
and its converted values to a date format handled by the
\code{\link{lubridate}} package.
}
\keyword{internal}

