\docType{classes}
\name{ListOfClasses}
\alias{ListOfClasses}
\alias{ListOfClasses-class}
\title{List of classes}
\description{
  An S4 subclass of \code{list} which requires all elements
  of the list to be the same class.
}
\section{Slots}{
  \describe{ \item{\code{.Data}}{Object as \code{list}.}
  \item{\code{classtype}}{Object as \code{character}.
  Required classtype for all elements in the list.} }
}

\section{Extends}{
  \describe{ \item{\code{list}}{Directly.} }
}
\keyword{classes}

