\name{subclass_list_of_classes}
\alias{subclass_list_of_classes}
\title{Create subclass list of classes}
\usage{
  subclass_list_of_classes(class, classtype = "ANY",
    where = topenv(parent.frame()))
}
\description{
  Creates a new subclass of \code{ListOfClasses} which
  requires a specific class.
}

